Introduce OrganizationReaderError
It's pretty common that someone can't assume a role they think they
should be able to assume only for it to turn out to be that their
original role can't assume OrganizationReader in order to FIND the role
they actually want to assume. This patch makes it abundantly clear when
they're in this situation so they can stop wondering and ultimately pull
less of their hair out.