Switch from stub providers to configuration_aliases
This is the main 0.14 to 0.15 upgrade step. It's not _strictly_
necessary as stub providers are deprecated not broken, but it looks a
lot more professional for us to be on the leading edge of this change.
Plus it's nice to see the terraform.Directory interface growing in a
reasonable direction.