Partially manage versions.tf
If a versions.tf file already exists, back off from managing the whole
file and instead only manage the versions of the providers Substrate
uses and Terraform itself. This allows folks to add more providers of
their own choosing.

This does not, however, allow folks to add their own additional
permutations of the AWS provider as it does not add partial management
of configuration_aliases.