Embed substrate-intranet.zip fully formed
This removes quite a bit of time from the build process and massively
reduces the complexity of the distribution by bundling the
cross-compiled bit into the binary that needs it. Furthermore, it stops
using the archive and external providers which means we can remove them
in the not too distant future.