PeeringConnections now preserves Add order
This specific-purpose set is now a specific-purpose sorted set, which
means that the order in which substrate-bootstrap-network-account
constructs peering connections is predictable, sensible, and stable, and
so is the output of substrate-root-modules.

This wasn't a problem in prior months because
substrate-bootstrap-network-account was processing during enumeration
such that the lack of order preservation in the map didn't matter.