Stop asking to configure a SAML provider
This opened quite a can of worms and identified some pretty gnarly bugs
that had snuck into a month or two's releases while I wasn't thinking
about onboarding as hard as I should've been.

Here we keep managing a SAML provider if one exists and, most
importantly, authorizing it to assume the roles folks would assume in
the AWS Console if they took the IdP route to get there.

If there isn't already a SAML provider, though, we don't create one. Now
that we have the Intranet logging folks into the AWS Console there's
really no reason to bother with the SAML route which everyone hates.