Support parameters in dispatch maps
The parameters can be a concrete or pointer type (with support for the
various reference types being somewhat easy to add) and any arity. The
types of every entry in the map must agree or the generated dispatch map
will not compile.

It cops out and relies on goimports to resolve import paths for the
arguments. It's fine.