Bug fix: Always get environment variables to Terraform
As it happened, I noted the correct fix the first time I encountered the
bug, wrote it down, and forgot about it. Then I encountered the bug for
a second time and described it ON THE LINE BELOW MY FIRST NOTE.

The ultimate problem was not setting environment variables on every path
through NewSession that could accept an access key interactively and
then return. As noted early on, setting environment variables when we
accept credentials interactively is the best solution.