Actually, we would want Wait to block here
If no one ever called Post, we would want Wait to wait forever. Well, we
wouldn't really but we do have to respect that's a generalization of
allowing Post and Wait to race with each other. It's possible for two
goroutines, one that eventually calls Post and one that eventually
calls Wait, to execute lopsidedly enough that the waiter gets to Wait
before the poster gets to Post. And that has to work.

This reverts commit ccb4a4f1a1d792c13453ba87194155048f1715fd.