Don't let telemetry failures scuttle the works
If telemetry was deferred (due to lack of credentials, most likely),
it'd be retried when we get some. Problem is, a failure there would fail
the whole program.

There were two failures here. The first is fixed in this commit: The
deadline was way too short, given how long it can take to get new IAM
credentials working. The second is fixed in the next commit.

In addition, this is a more idiomatic use of ctx.