Fix the environment variable clobbering bug
I thought it was only going to be problematic in Lambda (and I didn't
even think of that at first) but it's actually problematic anyplace we
use SetCredentials. That was a lot more places before finishing the
v1-to-v2 migration, which used SetCredentials in all the commands that
needed to dual-stack, however briefly.