Truly don't read ~/.aws/{config,credentials}
This has always been what I've been trying to do. I even thought I was.
I was wrong and proven so when someone set

    [profile default]
    credential_process = substrate credentials -format json -quiet

and created an infinite recursion of opening browser tabs. When I tested
this I had to restart my computer.

The reason was that these nil slices were instructing the AWS SDK to do
its default thing, not to do nothing, so when the default profile ran
`substrate credentials -format json -quiet`, that process loaded the
default profile which ran the command again, etc.