Bug fix: Hackily fix bad interaction with -quality and others
Since -quality can have a default, it was triggering this invalid
argument test when ideally it shouldn't because the user didn't actually
provide that value.

This fix is a hacky shortcut but is sufficient because the odds of
someone actually supplying just -quality and either -management or
-special and needing to get an error from that are shockingly low. In
every imaginable case they'll have also provided -domain and
-environment and thus been given the expected error, regardless of
what's going on with -quality.