Factor working directory management into main
This is necessary because QualityFlag needs to see find files in the
Substrate repository, so the cmdutil.Chdir needs to come before flags
are defined and parsed.

Then there are commands that weren't managing the working directory at
all that now are because it's factored into main. This necessitated
introducing UndoChdir and a package-global variable, which is
distasteful but could be much worse.