Grab substrate-intranet.zip every Terraform run
I have a theory that this conditional was the cause of a customer
experiencing Intranet downgrades. Specifically, I suspect someone was
running Terraform directly on an old work tree after someone else had
performed a Substrate upgrade.

I still need to confirm that a downgrade here will explode as it should.