Support -format json in intranet-zip
This variation, when combined with -base64sha256, will write a JSON
object that can be sensibly extracted from Terraform
data.external.zip.result objects.

This extra function is necessary because the filebase64sha256 function
happens too early in a Terraform run to see the content placed by the
external command.