Allow `terraform plan` to fail
Sometimes, as when adding a region and one of your modules has added
additional variables not accounted for in the generated Terraform code,
Terraform won't be able to plan a root module. That's OK, though! That's
data! It's pretty hard to work with, though, when it aborts the whole
process because it never even generates the code for root modules it
hasn't reached yet.

This way, we allow `terraform plan` to exit non-zero, though we still
fail hard if e.g. terraform isn't found on PATH.