Add a Version field to accounts tables
The version of Substrate that last managed an account is stored in the
SubstrateVersion tag on each account. This tag is used to detect and
prevent (or warn about, for some commands) Substrate downgrades to help
teams get their upgrades propagated. It's done per account because
upgrades are not atomic across a whole AWS organization.

No one asked for this but it seems like a nice thing that people who are
doing Substrate upgrades will appreciate _and_ it seems like a
reasonable opportunity to reacquaint myself with GitHub pull requests,
which I haven't used in genuinely three years.