Toggle IMDSv2 enforcement via a file
IMDSv2 trips a lot of Substrate users up and stands as a potentially
serious impediment to mature organizations that haven't gotten around to
enforcing the use of IMDSv2 adopting or even trying out Substrate.

This patch allows rudimentary toggling of the IMDSv2-related statements
in the Substrate-managed SCP so that organizations can opt-out.