Extract the account's tags from the filesystem
The upgrade to Terraform 1.3.6 (probably) introduced a bug in which new
accounts couldn't complete their initial Terraform run. The error
message is as follows:

    │ Error: Invalid count argument
    │
    │   on ../../../../modules/substrate/regional/main.tf line 4, in data "aws_subnets" "private":
    │    4:   count = module.global.tags.environment == "admin" ? 0 : 1
    │
    │ The "count" value depends on resource attributes that cannot be determined until apply, so Terraform cannot predict how many instances will be created. To
    │ work around this, use the -target argument to first apply only the resources that the count depends on.

The problem is that modules/substrate/regional (one of the modules
completely generated by Substrate) uses the output of
modules/substrate/global (which it encapsulates) in a way that must have
become unsupported in Terraform 1.3. (I'm not bothering to look up when
this broke because it's frankly irrelevant.) Specifically, counts are
evaluated before data sources and it's a data source that's behind
modules.global.tags.environment in the error message above.

We can't remove the count because modules/substrate/regional truly must
parameterize by environment in order to identify all the subnets but
also not go looking for any that don't exist. (Terraform is pretty
unforgiving in this respect.)

So we have to find a way to make modules/substrate/global not use a data
source. Mercifully, Substrate's filesystem layout gives us a way out of
this mess. (The parse is a little more complicated than one might expect
to accommodate the unlikely but not impossible scenario in which
"root-modules" appears twice in the absolute pathname to a root module.

This has been tested by creating an admin and a service account plus
recreating lots of preexisting accounts.