It'd be cool to have a "Var" version of StringSliceFlag but this ain't it
It's not possible to provide a slice and then let something extend that
slice while holding onto the original address since the slice returned
by append() will not (with near certainty in case of an originally empty
slice) have the same address.

Oh well. If we do end up needing this, the sad answer is to have the
caller provide a *StringSlice instead of a []string.