Failing tests showing why we need -all-domains, etc.
If you run with just e.g. -management, the sensible expectation is that
it only creates the role in the management account. However, without any
-domain, -environment, and -quality arguments, it assumes it should
create the role in all serivce accounts. That makes sense when there's
not also a -management or similar but is confusing when there is. The
result is that there's no way to create a role _just_ in the management
or one or more special accounts.

These tests show this clearly and the solution (coming soon) will be to
require folks to explicitly opt into roles being created in all service
accounts by providing some combination of -all-domains,
-all-environments, and -all-qualities.