Make testawscfg also change directories
The error that was being ignored was telling us that substrate.qualities
couldn't be found and it couldn't because we weren't in the test1
directory when we tried to assume a role in its organization.

This has to be done with the working directory and not by setting
SUBSTRATE_ROOT because the latter is handled in main.main which is not
under test.