Fix -assume-role-policy "" and -policy "" bug
The problem is that we weren't detecting when these options weren't
given at all. In that case, the tag that stores their value isn't set
and so becomes "" when we pull it out of the map. Splitting "" doesn't
result in []string{}, it results in []string{""}.

Now we're actually checking whether there was a tag and that it had a
non-empty value.