Cache an access key in Secrets Manager
This commit makes Credential Factory a lot faster all but about once a
week by caching an access key for the CredentialFactory IAM user in
Secrets Manager, accessible only to the Intranet role.

Before trying to create an access key we check for one in Secrets
Manager. If it was created less than a week ago, we use that one. If
it's expired, we delete it and fall through.

Next we do exactly what we did before, trying to create an access key a
few times since it's a race for a scarce resource. If we're able to
create one, we cache it in Secrets Manager.