Write required versions to the Substrate repo if they don't exist
This is both the required Terraform version and the minimum version
constraint for the Terraform AWS provider.

This is step one in decoupling Terraform upgrades from Substrate
upgrades. We're bootstrapping what will from now on be a
customer-directed upgrade cycle. This is very much enabled by
Substrate's goal of generating less (to the point of no) Terraform code
over time. It's highly unlikely they're going to make it too
incompatible in the meantime to make this hard.

We can't upgrade the Terraform AWS provider all the way to the latest
version, which is 5.14.0 at the time of this commit, because they
removed the deprecated aws_subnet_ids data source which we use in
network peering and which is impossible to refactor out of while staying
within Terraform. We'll get out of this and unblock these upgrades when
we move network peering into Go.