Support regions.Default not appearing in regions.Selected
This bug plagued a prospective customer for a few days. They'd
configured us-east-1 as their default region and us-west-2 as their only
service account region. This is supposed to be supported. Instead, they
were greeted by Terraform errors saying a bucket doesn't exist on the
first global root module Substrate tried to init/plan/apply on their
behalf.

The solution is to explicitly manage S3 and DynamoDB resources for
remote Terraform state in both the default region and every other
selected region. We take a bit of care here to de-dupe the list so that
we're not unnecessarily making things slower and making the output more
confusing.