Bug fix: Don't leak AccessToken across requests
If we keep an AccessToken in an oauthoidc.Client, now that the
authorizer and the actual Lambda functions are one and the same, then we
end up using some AccessToken for a pre-auth request that really needs
to be sent with just the ClientId and clientSecret. By copying the
oauthoidc.Client more proactively, we avoid this.