Create the src-bin AWS config before changing directories
This really sent me down a rabbit hole for a situation that simply
cannot happen outside of the test suite. Still, the local cache of
DescribeOrganization output is proving to be a little dangerous because
there's no way to detect when we get an incorrect result if we change
directories, hit _that_ cache, and then carry on. And once we assume a
role, the cache hit _becomes_ valid.