Move debug output around to be more clear and more useful
We couldn't leave it where it was because the bug fix (revert-ish)
coming in the next commit would make the debug output useless again, as
it was at the time that prompted me to stupidly introduce the bug. But
there's no good place to put it without some redundancy, so I decided to
embrace the redundancy and label each policy the debug flag prints with
"creating" or "updating". In most cases, it'll only get to the
"updating" side but on a first run it'll print both and the debug output
will clearly show Substrate solving the chicken-and-egg problem of AWS
IAM assume-role policies.